+++ src/ray/common/network_util.cc
@@ -18,13 +18,3 @@
 #include "ray/util/logging.h"
 
 using boost::asio::ip::tcp;
-
-bool CheckPortFree(int port) {
-  instrumented_io_context io_service;
-  tcp::socket socket(io_service);
-  socket.open(boost::asio::ip::tcp::v4());
-  boost::system::error_code ec;
-  socket.bind(boost::asio::ip::tcp::endpoint(boost::asio::ip::tcp::v4(), port), ec);
-  socket.close();
-  return !ec.failed();
-}
