+++ src/ray/common/test/BUILD.bazel
@@ -146,19 +146,6 @@ ray_cc_test(
     ],
 )
 
-ray_cc_test(
-    name = "client_connection_test",
-    size = "small",
-    srcs = ["client_connection_test.cc"],
-    tags = ["team:core"],
-    deps = [
-        "//src/ray/common:asio",
-        "//src/ray/common:id",
-        "//src/ray/common:network",
-        "@com_google_googletest//:gtest_main",
-    ],
-)
-
 ray_cc_library(
     name = "testing",
     testonly = True,
