+++ src/ray/ipc/BUILD.bazel
@@ -1,0 +1,18 @@
+load("//bazel:ray.bzl", "ray_cc_library")
+
+ray_cc_library(
+    name = "client_connection",
+    srcs = [
+        "client_connection.cc",
+    ],
+    hdrs = [
+        "client_connection.h",
+    ],
+    deps = [
+        "//src/ray/common:asio",
+        "//src/ray/common:event_stats",
+        "//src/ray/common:id",
+        "//src/ray/common:status",
+        "//:node_manager_fbs",
+    ],
+)
