+++ src/ray/ipc/client_connection.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "ray/common/client_connection.h"
+#include "ray/ipc/client_connection.h"
 
 #include <boost/asio/buffer.hpp>
 #include <boost/asio/generic/stream_protocol.hpp>
