+++ src/ray/ipc/test/BUILD.bazel
@@ -1,0 +1,15 @@
+load("//bazel:ray.bzl", "ray_cc_test")
+
+ray_cc_test(
+    name = "client_connection_test",
+    size = "small",
+    srcs = ["client_connection_test.cc"],
+    tags = ["team:core"],
+    deps = [
+        "//src/ray/common:asio",
+        "//src/ray/common:id",
+        "//src/ray/ipc:client_connection",
+        "@boost//:asio",
+        "@com_google_googletest//:gtest_main",
+    ],
+)
