+++ src/ray/object_manager/plasma/BUILD.bazel
@@ -103,7 +103,7 @@ ray_cc_library(
         "//src/ray/common:asio",
         "//src/ray/common:file_system_monitor",
         "//src/ray/common:id",
-        "//src/ray/common:network",
+        "//src/ray/ipc:client_connection",
         "//src/ray/common:ray_config",
         "//src/ray/common:status",
         "//src/ray/object_manager:object_manager_common",
@@ -280,7 +280,7 @@ ray_cc_library(
         ":object_manager_plasma_common",
         ":plasma_generated",
         "//src/ray/common:id",
-        "//src/ray/common:network",
+        "//src/ray/ipc:client_connection",
         "//src/ray/common:status",
         "//src/ray/object_manager:object_manager_common",
         "//src/ray/protobuf:common_cc_proto",
