+++ src/ray/raylet_client/raylet_client.h
@@ -24,7 +24,7 @@
 #include "ray/common/asio/instrumented_io_context.h"
 #include "ray/common/buffer.h"
 #include "ray/common/bundle_spec.h"
-#include "ray/common/client_connection.h"
+#include "ray/ipc/client_connection.h"
 #include "ray/common/status.h"
 #include "ray/common/status_or.h"
 #include "ray/common/task/task_spec.h"
