+++ src/ray/raylet_client/raylet_connection.h
@@ -21,7 +21,7 @@
 #include <vector>
 
 #include "ray/common/asio/instrumented_io_context.h"
-#include "ray/common/client_connection.h"
+#include "ray/ipc/client_connection.h"
 
 namespace ray::raylet {
